tests/wpt/tests/css/css-mixins/mixins/mixin-basic.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Mixins: Basic test</title>
    <link rel="help" href="https://drafts.csswg.org/css-mixins-1/#defining-mixins">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
      @mixin --m1() {  /* Will be overwritten. */
        @result {
          .cls {
            color: red;
          }
        }
      }
      @mixin --m1() {
        @result {
          .cls {
            color: green;
          }
        }
      }
      @mixin invalid-name() {
        @result {
          .cls {
            color: red;
          }
        }
      }
      @mixin --no-parameter-list {
        @result {
          .cls {
            border: 1px solid green;
          }
        }
      }
      @mixin --missing-result1() {
       .cls {
          color: red;
        }
      }
      @mixin --missing-result2() {
        color: red;
      }
      @mixin --empty1() {}
      @mixin --empty2() { @result {} }
      div {
        @apply --m1;
        @apply invalid-name;
        @apply --no-parameter-list;
        @apply --missing-result1;
        @apply --missing-result2;
        @apply --empty1;
        @apply --empty2;
      }
    </style>
  </head>
  <body>
    <div><div class="cls" id="target">This text should be green with a green border.</div></div>
  <script>
    test(() => {
      let target = document.getElementById('target');
      assert_equals(getComputedStyle(target).color, 'rgb(0, 128, 0)');
      assert_equals(getComputedStyle(target).borderTopColor, 'rgb(0, 128, 0)');
      assert_equals(getComputedStyle(target).borderTopWidth, '1px');
      assert_equals(getComputedStyle(target).borderTopStyle, 'solid');
    });
  </script>
  </body>
</html>
